pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ivy"
version = "0.1.0"
description = "Assistant local IVY"
authors = [{name = "IVY"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "apscheduler",
    "llama-cpp-python",
    "numpy",
    "faiss-cpu",
    "sentence-transformers",
    "pytesseract",
    "python-docx",
    "openpyxl",
    "python-pptx",
    "pyzipper",
    "prometheus-client",
    "psutil",
    "tzdata",
    "python-dotenv",
    "pdf2image",
    "pillow",
    "duckduckgo-search>=6.1.4",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "itsdangerous",
    "typer[all]",
    "watchdog",
    "aiosqlite",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "isort",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "playwright",
    "pip-audit",
    "bandit",
]
voice = [
    "piper-tts==1.3.0",
    "onnxruntime==1.20.1",
    "faster-whisper==1.2.0",
    "ctranslate2==4.6.0",
    "sounddevice==0.5.3",
    "webrtcvad==2.0.10",
    "PySide6==6.10.0",
    "pywin32>=306; platform_system == \"Windows\"",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = [
  "app",
  "core",
  "plugins",
  "llama_cpp",
]
